+++ swarms-rs/src/structs/mod.rs
@@ -3,3 +3,6 @@ pub mod conversation;
 pub mod graph_workflow;
 pub mod persistence;
 pub mod tool;
+
+#[cfg(test)]
+mod test_utils;
